% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_dataset.R
\name{fetch_dataset}
\alias{fetch_dataset}
\title{Download dataset from API}
\usage{
fetch_dataset(
  wh_dataset = "caiso_fuel_mix",
  start_time = "2024-04-06",
  end_time = "2024-04-11",
  api_key = Sys.getenv("GRIDSTATUS_API_KEY")
)
}
\arguments{
\item{wh_dataset}{Name of dataset to download}

\item{start_time}{Start time}

\item{end_time}{End time}

\item{api_key}{API key}
}
\value{
df: Dataframe fo requested dataset
}
\description{
Download dataset from API
}
