% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api_request.R
\name{get_api_request}
\alias{get_api_request}
\title{Make API GET request and return data}
\usage{
get_api_request(query_url, api_key = Sys.getenv("GRIDSTATUS_API_KEY"))
}
\arguments{
\item{query_url}{URL for the API query}

\item{api_key}{API key; defaults to key saved in renviron file}
}
\value{
data from the API response
}
\description{
Make API GET request and return data
}
