% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_source_names.R
\name{get_source_names}
\alias{get_source_names}
\title{Get unique data source names}
\usage{
get_source_names(info)
}
\arguments{
\item{info}{Output of get_available_datasets()}
}
\value{
Unique source names (ie ercot_fuel_mix >> ercot)
}
\description{
Get unique data source names
}
