% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_gen_long.R
\name{pivot_gen_long}
\alias{pivot_gen_long}
\title{Pivot generation (fuel-mix) data to long format}
\usage{
pivot_gen_long(df)
}
\arguments{
\item{df}{Dataframe of generation data in wide format (as returned by fetch_dataset())}
}
\value{
Dataframe pivoted to long format
}
\description{
Pivot generation (fuel-mix) data to long format
}
\seealso{
\code{\link[=fetch_dataset]{fetch_dataset()}}
}
