% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgridstatus-package.R
\docType{package}
\name{rgridstatus-package}
\alias{rgridstatus}
\alias{rgridstatus-package}
\title{rgridstatus: API Wrapper For GridStatus}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andypicke/rgridstatus}
  \item Report bugs at \url{https://github.com/andypicke/rgridstatus/issues}
}

}
\author{
\strong{Maintainer}: Andy Pickering \email{andypicke@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
