% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_query_url.R
\name{construct_query_url}
\alias{construct_query_url}
\title{Build url for API request}
\usage{
construct_query_url(
  wh_dataset = "caiso_fuel_mix",
  start_time = Sys.Date() - 5,
  end_time = Sys.Date()
)
}
\arguments{
\item{wh_dataset}{Name of dataset to download}

\item{start_time}{Start time. Default is 5 days ago.}

\item{end_time}{End time. Default is today's date.}
}
\value{
url
}
\description{
Build url for API request
}
