% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_dataset.R
\name{fetch_dataset}
\alias{fetch_dataset}
\title{Download dataset from API}
\usage{
fetch_dataset(
  wh_dataset = "caiso_fuel_mix",
  start_time = Sys.Date() - 5,
  end_time = Sys.Date(),
  api_key = Sys.getenv("GRIDSTATUS_API_KEY"),
  tz_local = "US/Pacific"
)
}
\arguments{
\item{wh_dataset}{Name of dataset to download. Default is "caiso_fuel_mix"}

\item{start_time}{Start time. Default is 5 days ago.}

\item{end_time}{End time. Default is today's date.}

\item{api_key}{API key. Default is "GRIDSTATUS_API_KEY" stored in .Renviron file}

\item{tz_local}{Local timezone to display times in column "datetime_local" (default "US/Pacific"). See available timezones with OlsonNames()}
}
\value{
df: Dataframe of requested dataset
}
\description{
Download dataset from API
}
